gitadora: address two-head review feedback
